refactor(stream): split 'superdesk-stream' module.
`superdesk-stream` becomes:
```
.
├── controllers
│   ├── StreamController.js
│   └── index.js
├── directives
│   ├── ActivityMessage.js
│   ├── ActivityStream.js
│   └── index.js
├── views
│   ├── activity-stream.html
│   └── workspace-stream.html
└── index.js

3 directories, 8 files
```
