+++ scripts/superdesk-stream/controllers/index.js
@@ -1,0 +1,1 @@
+export { StreamController } from './StreamController';
