+++ scripts/superdesk-stream/directives/index.js
@@ -1,0 +1,2 @@
+export { ActivityStream } from './ActivityStream';
+export { ActivityMessage } from './ActivityMessage';
